Improve ip address handling for master backend
If the backend server address in settings is set to an address that is
inaccessible to leanfront but the backend can be accessed throuigh a
different ip address or DNS entry, leanfront would fail to download
artwork or play videos.

For example a port-forward setup to allow accessing the backend from
the internet.

To fix this, it now uses the leaanfront settings ip address and port when
accessing the priomary backend for all purposes.

Slave backends are still accessed using the ip address from mythbackend
settings, so this ip address must be accessible.

If "Master Backend Override" is specified, recordings from slave
backends will work even if the slave ip address is inaccessible.